pyproject.toml
[tool.poetry]
name = "mais-todos"
version = "0.1.0"
description = ""
authors = ["robson"]
readme = "README.md"
packages = [{include = "mais_todos"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
djangorestframework = "^3.14.0"
django-cryptography = "^1.1"
psycopg2-binary = "2.9.5"
python-decouple = "^3.8"
creditcard = "^1.0.2"
dockerignore-generate = "^1.0.7"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.1.1"
pytest = "^7.2.2"
black = "^23.1.0"
flake8 = "^6.0.0"
mypy = "^1.1.1"
pycodestyle = "^2.10.0"
pytest-django = "^4.5.2"
isort = "^5.12.0"

[tool.isort]
force_single_line = true
line_length = 120
skip_glob = "**/migrations/*.py"
use_parentheses = true
multi_line_output = 5
include_trailing_comma = true
lines_between_types = 1
known_django = "django"
known_mais_todos = "mais_todos"
known_third_party = ["urllib3", "rest_framework", "request_id_django_log", "pytest"]
sections = ["FUTURE", "STDLIB", "DJANGO", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER", "MAIS_TODOS"]


[tool.black]
line-length = 120
target_version = ['py39']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | migrations
  )/
)
'''

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
